requirements.txt
browser_cookie3
numpy
pandas
pyktok
webvtt-py
papermill
moviepy
azure-cognitiveservices-speech
python-dotenv